[Bronze IV] Title: Congruent Numbers, Time: 64 ms, Memory: 11568 KB -BaekjoonHub

+++ 백준/Bronze/15128. Congruent Numbers/Congruent Numbers.java
@@ -5,10 +5,10 @@ public class Main {
     public static void main(String[] args) throws Exception {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         StringTokenizer st = new StringTokenizer(br.readLine());
-        double p1 = Double.parseDouble(st.nextToken());
-        double q1 = Double.parseDouble(st.nextToken());
-        double p2 = Double.parseDouble(st.nextToken());
-        double q2 = Double.parseDouble(st.nextToken());
+        long p1 = Long.parseLong(st.nextToken());
+        long q1 = Long.parseLong(st.nextToken());
+        long p2 = Long.parseLong(st.nextToken());
+        long q2 = Long.parseLong(st.nextToken());
         System.out.print((p1 * p2) % (q1 * q2 * 2) == 0 ? 1 : 0);
     }
 }
